# le-backend/docs/openapi_enhanced.yaml
openapi: 3.0.3
info:
  title: LC Workflow API - Enhanced
  description: |
    Enhanced LC Workflow API with comprehensive system stability improvements, 
    security features, monitoring capabilities, and data synchronization.
    
    ## Features
    
    - **Enhanced File Management**: Secure file upload with malware scanning and encryption
    - **Folder Organization**: Automatic document organization by type
    - **Health Monitoring**: Comprehensive system health checks and metrics
    - **Data Synchronization**: Real-time updates and cache management
    - **Security**: Advanced authentication, authorization, and audit logging
    - **Feature Flags**: Gradual rollout and A/B testing capabilities
    
    ## Authentication
    
    All endpoints require JWT authentication unless specified otherwise.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Error Handling
    
    The API uses standardized error responses with correlation IDs for tracking.
    All errors include suggested actions and detailed information.
    
    ## Rate Limiting
    
    - File Upload: 100 requests/hour per user
    - General API: 1000 requests/hour per user
    - Health Checks: 60 requests/minute (no user limit)
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /feature-flags/:
    get:
      tags:
        - Feature Flags
      summary: Get all feature flags
      description: Get all feature flags (admin only)
      operationId: getAllFeatureFlags
      parameters:
        - name: enabled_only
          in: query
          schema:
            type: boolean
            default: false
          description: Return only enabled flags
        - name: flag_type
          in: query
          schema:
            type: string
          description: Filter by flag type
      responses:
        '200':
          description: Feature flags retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureFlagResponse'

    post:
      tags:
        - Feature Flags
      summary: Create feature flag
      description: Create a new feature flag (admin only)
      operationId: createFeatureFlag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureFlagRequest'
      responses:
        '201':
          description: Feature flag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagResponse'
  /feature-flags/{flag_name}:
    get:
      tags:
        - Feature Flags
      summary: Get feature flag
      description: Get a specific feature flag (admin only)
      operationId: getFeatureFlag
      parameters:
        - name: flag_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Feature flag retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagResponse'

    put:
      tags:
        - Feature Flags
      summary: Update feature flag
      description: Update an existing feature flag (admin only)
      operationId: updateFeatureFlag
      parameters:
        - name: flag_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureFlagUpdateRequest'
      responses:
        '200':
          description: Feature flag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagResponse'
  /feature-flags/{flag_name}/check:
    get:
      tags:
        - Feature Flags
      summary: Check feature flag
      description: Check if a feature flag is enabled for a user
      operationId: checkFeatureFlag
      parameters:
        - name: flag_name
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
          description: User ID to check flag for
      responses:
        '200':
          description: Flag status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  flag_name:
                    type: string
                  user_id:
                    type: string
                  user_role:
                    type: string
                  enabled:
                    type: boolean
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FeatureFlagRequest:
      type: object
      required:
        - name
        - description
        - flag_type
        - rollout_strategy
      properties:
        name:
          type: string
        description:
          type: string
        flag_type:
          type: string
          enum: [boolean, percentage, user_list, role_based, date_range]
        enabled:
          type: boolean
          default: false
        rollout_strategy:
          type: string
          enum: [all_users, percentage_based, user_whitelist, role_based, gradual_rollout]
        rollout_percentage:
          type: number
          minimum: 0
          maximum: 100
          default: 0
        user_whitelist:
          type: array
          items:
            type: string
        role_whitelist:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date-time
          nullable: true
        end_date:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
    FeatureFlagUpdateRequest:
      type: object
      properties:
        description:
          type: string
        enabled:
          type: boolean
        rollout_strategy:
          type: string
        rollout_percentage:
          type: number
          minimum: 0
          maximum: 100
        user_whitelist:
          type: array
          items:
            type: string
        role_whitelist:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date-time
          nullable: true
        end_date:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
    FeatureFlagResponse:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        flag_type:
          type: string
        enabled:
          type: boolean
        rollout_strategy:
          type: string
        rollout_percentage:
          type: number
        user_whitelist:
          type: array
          items:
            type: string
        role_whitelist:
          type: array
          items:
            type: string
        start_date:
          type: string
          format: date-time
          nullable: true
        end_date:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Error Response Schema
...